uwb_nlos/requirements.txt
# ── Development machine (training) ────────────────────────────
# Python 3.11 / 3.12 recommended for full torch + onnxruntime support
numpy>=1.26
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
matplotlib>=3.7
torch>=2.2           # CPU-only: pip install torch --index-url https://download.pytorch.org/whl/cpu
onnx>=1.15
onnxruntime>=1.17
pyserial>=3.5
